refactor: 优化DraggableItem 为函数组件

+++ src/views/tool/build/DraggableItem.vue
@@ -1,19 +1,20 @@
 <script>
 import draggable from 'vuedraggable'
 import render from '@/utils/generator/render'
+import Vue from 'vue'
 
 const components = {
   itemBtns(h, element, index, parent) {
-    const { copyItem, deleteItem } = this.$listeners
+    const { copyItem, deleteItem } = this.listeners
     return [
       <span class="drawing-item-copy" title="复制" onClick={event => {
-        copyItem(element, parent)
+        copyItem && copyItem(element, parent)
         event.stopPropagation()
       }}>
         <i class="el-icon-copy-document"/>
       </span>,
       <span class="drawing-item-delete" title="删除" onClick={event => {
-        deleteItem(index, parent)
+        deleteItem && deleteItem(index, parent)
         event.stopPropagation()
       }}>
         <i class="el-icon-delete"/>
@@ -24,9 +25,9 @@ const components = {
 /* eslint-disable indent */
 const layouts = {
   colFormItem(h, element, index, parent) {
-    const { activeItem } = this.$listeners
+    const { activeItem } = this.listeners
     let className = this.activeId === element.formId ? 'drawing-item active-from-item' : 'drawing-item'
-    if (this.formConf.unFocusedComponentBorder) className += ' unfocus-bordered'
+    if (this.props.formConf.unFocusedComponentBorder) className += ' unfocus-bordered'
     return (
         <el-col span={element.span} class={className}
                 nativeOnClick={event => {
@@ -36,7 +37,7 @@ const layouts = {
           <el-form-item label-width={element.labelWidth ? `${element.labelWidth}px` : null}
                         label={element.label} required={element.required}>
             <render key={element.renderKey} conf={element} onInput={event => {
-              this.$set(element, 'defaultValue', event)
+              Vue.set(element, 'defaultValue', event)
             }}/>
           </el-form-item>
           {components.itemBtns.apply(this, arguments)}
@@ -44,8 +45,8 @@ const layouts = {
     )
   },
   rowFormItem(h, element, index, parent) {
-    const { activeItem } = this.$listeners
-    const className = this.activeId === element.formId ? 'drawing-row-item active-from-item' : 'drawing-row-item'
+    const { activeItem } = this.listeners
+    const className = this.props.activeId === element.formId ? 'drawing-row-item active-from-item' : 'drawing-row-item'
     let child = renderChildren.apply(this, arguments)
     if (element.type === 'flex') {
       child = <el-row type={element.type} justify={element.justify} align={element.align}>
@@ -86,10 +87,7 @@ function layoutIsNotFound() {
 }
 
 export default {
-  components: {
-    render,
-    draggable
-  },
+  functional: true,
   props: [
     'element',
     'index',
@@ -97,11 +95,10 @@ export default {
     'activeId',
     'formConf'
   ],
-  render(h) {
-    const layout = layouts[this.element.layout]
-
+  render(h, context) {
+    const layout = layouts[context.props.element.layout]
     if (layout) {
-      return layout.call(this, h, this.element, this.index, this.drawingList)
+      return layout.call(context, h, context.props.element, context.props.index, context.props.drawingList)
     }
     return layoutIsNotFound()
   }
